Bien hecho spring: controllers

+++ PSP/SpringApiRest/src/main/java/org/example/springdemo/domain/model/Entrenamiento.java
@@ -1,8 +1,8 @@
 package org.example.springdemo.domain.model;
 
 public record Entrenamiento(
-        int id,
-        int usuarioId,
+        Long id,
+        Integer usuarioId,
         String nombre,
         String descripcion) {
 }
